Fixes aiohttp_mock_app in director's tests
